convert tests to mocha, add to prepublish
