+++ .eslintignore
@@ -1,2 +1,1 @@
-**/*.tests.js
 demos/**
