--- builder.tests.js
@@ -1,45 +1,0 @@
-﻿define(["knockout", "./builder",
-	"mock-ajax"],
-	function (ko, ViewModel) {
-
-		describe("Address Builder ViewModel", function () {
-
-
-			var data = {
-				address1: ko.observable("950 Gravier"),
-				address2: ko.observable("1700"),
-				city: ko.observable("New Orleans"),
-				state: ko.observable("LA"),
-				postalCode: ko.observable("70123"),
-				country: ko.observable("United States"),
-				instanceName: "addy-1"
-			};
-
-			var model;
-
-			describe("When creating a new model with address data", function () {
-				beforeEach(function () {
-					model = new ViewModel({ address: data, instanceName: data.instanceName });
-				}, this);
-
-				it("should be populated with an address of observable values", function () {
-					expect(model.address().address1()).toEqual("950 Gravier");
-					expect(model.address().address2()).toEqual("1700");
-					expect(model.address().city()).toEqual("New Orleans");
-					expect(model.address().state()).toEqual("LA");
-					expect(model.address().postalCode()).toEqual("70123");
-					expect(model.address().country()).toEqual("US");
-				}, this);
-
-				it("should have countries", function () {
-					expect(model.countries().length > 0).toEqual(true);
-				});
-
-				it("should have states", function () {
-					expect(model.states().length > 0).toEqual(true);
-				});
-
-			});
-
-		});
-	});
