+++ package.json
@@ -4,8 +4,9 @@
   "description": "address bindings and extension for observable address objects",
   "main": "knockout.address.js",
   "scripts": {
+    "test": "mocha",
     "lint": "eslint . --format=unix",
-    "prepublish": "npm run lint",
+    "prepublish": "npm run lint && mocha",
     "devboiler": "grunt",
     "build": "webpack --progress && npm run devboiler",
     "watch": "webpack --progress --watch",
@@ -31,14 +32,16 @@
   "devDependencies": {
     "babel-eslint": "6.x",
     "bootstrap": "3.x",
+    "chai": "3.x",
     "eslint": "2.x",
     "eslint-plugin-filenames": "1.x",
     "grunt": "1.x",
     "grunt-cli": "1.x",
     "grunt-contrib-copy": "1.x",
-    "http-server": "^0.9.0",
+    "http-server": "0.9.x",
     "jquery": "^3.0.0",
     "jquery-extend": "2.x",
+    "mocha": "2.x",
     "webpack": "1.x"
   },
   "dependencies": {
