+++ test/.eslintrc.yml
@@ -1,0 +1,2 @@
+env:
+  mocha: true
